переделал передачу query params

+++ project/libs/blog/blog-like/src/blog-like-module/blog-like.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { BlogLikeRepository } from './blog-like.repository';
 import { BlogLikeEntity } from './blog-like.entity';
 
@@ -21,7 +21,11 @@ export class BlogLikeService {
     return await this.blogLikeRepository.findById(id);
   }
 
-  public async remove(id: string): Promise<void> {
+  public async remove(id: string, userId: string): Promise<void> {
+    const like = await this.findById(id);
+    if (like.userId !== userId) {
+      throw new HttpException('Невозможно удалить лайк другого пользователя', HttpStatus.BAD_REQUEST);
+    }
     return await this.blogLikeRepository.deleteById(id);
   }
 }
